Большой пак изменений
рефакторинг StarSystemForm
Добавлена рабочая система захвата
