+++ .gitignore
@@ -38,3 +38,5 @@ GalaxyConquest/obj/x86/Debug/GalaxyConquest.Tech_Tree.resources
 GalaxyConquest/obj/x86/Debug/TempPE/Properties.Resources.Designer.cs.dll
 GalaxyConquest/bin/Debug/teches.xml
 GalaxyConquest/bin/Debug/Starnames backup.xml
+GalaxyConquest/Sprites/background/Thumbs.db
+GalaxyConquest/Resources/Thumbs.db
