+++ GalaxyConquest/SpaceObjects/Planet.cs
