+++ GalaxyConquest/SpaceObjects/PlanetForm.resx
