+++ GalaxyConquest/SpaceObjects/SpaceObject.cs
