+++ GalaxyConquest/SpaceObjects/StarWarp.cs
@@ -1,0 +1,18 @@
+﻿using System;
+
+namespace GalaxyConquest
+{
+    [Serializable]
+    public class StarWarp : SpaceObject
+    {
+        public int type;
+
+        public StarSystem system1;
+        public StarSystem system2;
+
+        public override void Move(double time)
+        {
+            return;
+        }
+    }
+}
