--- GalaxyConquest/StarWarp.cs
@@ -1,14 +1,0 @@
-﻿using System;
-
-namespace GalaxyConquest
-{
-    [Serializable]
-    public class StarWarp
-    {
-        public string name;
-        public int type;
-
-        public StarSystem system1;
-        public StarSystem system2;
-    }
-}
