+++ GalaxyConquest/Tech_Tree.cs
@@ -5,6 +5,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Xml.Linq;
 using System.Xml.Serialization;
+using GalaxyConquest.Game;
 
 namespace GalaxyConquest
 {
